[NF] bot core updated
