+++ src/main/java/com/botmasterzzz/social/service/MessageProcess.java
@@ -1,0 +1,9 @@
+package com.botmasterzzz.social.service;
+
+import com.botmasterzzz.bot.api.impl.objects.OutgoingMessage;
+
+public interface MessageProcess {
+
+    void process(OutgoingMessage apiMethod, Long kafkaKey);
+
+}
